requirements.txt
# Core Django and REST Framework
requests
gunicorn
psutil
python-decouple
Django==4.2.16
django-bootstrap==0.2.4
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

